AWS amplify added, not yet fully integrated
