+++ amplify/.config/project-config.json
@@ -1,0 +1,13 @@
+{
+  "projectName": "taskmaster",
+  "version": "3.1",
+  "frontend": "android",
+  "providers": [
+    "awscloudformation"
+  ],
+  "android": {
+    "config": {
+      "ResDir": "app/src/main/res"
+    }
+  }
+}
