+++ amplify/backend/api/amplifyDatasource/parameters.json
@@ -1,0 +1,5 @@
+{
+  "AppSyncApiName": "amplifyDatasource",
+  "DynamoDBBillingMode": "PAY_PER_REQUEST",
+  "DynamoDBEnableServerSideEncryption": "false"
+}
