+++ amplify/backend/api/amplifyDatasource/schema.graphql
@@ -1,0 +1,14 @@
+type Task @model @auth(rules: [{allow: public}]){
+  id: ID!
+  title: String!
+  body: String
+  taskState: State
+}
+
+enum State {
+  NEW
+  ASSIGNED
+  COMPLETE
+}
+
+
