+++ amplify/backend/api/amplifyDatasource/transform.conf.json
@@ -1,0 +1,9 @@
+{
+    "Version": 5,
+    "ResolverConfig": {
+        "project": {
+            "ConflictHandler": "AUTOMERGE",
+            "ConflictDetection": "VERSION"
+        }
+    }
+}
