+++ amplify/backend/types/amplify-dependent-resources-ref.d.ts
@@ -1,0 +1,9 @@
+export type AmplifyDependentResourcesAttributes = {
+    "api": {
+        "amplifyDatasource": {
+            "GraphQLAPIKeyOutput": "string",
+            "GraphQLAPIIdOutput": "string",
+            "GraphQLAPIEndpointOutput": "string"
+        }
+    }
+}
