+++ amplify/team-provider-info.json
@@ -1,0 +1,15 @@
+{
+  "dev": {
+    "awscloudformation": {
+      "AuthRoleName": "amplify-taskmaster-dev-201511-authRole",
+      "UnauthRoleArn": "arn:aws:iam::573887199718:role/amplify-taskmaster-dev-201511-unauthRole",
+      "AuthRoleArn": "arn:aws:iam::573887199718:role/amplify-taskmaster-dev-201511-authRole",
+      "Region": "us-east-1",
+      "DeploymentBucketName": "amplify-taskmaster-dev-201511-deployment",
+      "UnauthRoleName": "amplify-taskmaster-dev-201511-unauthRole",
+      "StackName": "amplify-taskmaster-dev-201511",
+      "StackId": "arn:aws:cloudformation:us-east-1:573887199718:stack/amplify-taskmaster-dev-201511/7825c130-afbe-11ec-ba48-0ec87615267b",
+      "AmplifyAppId": "dv3dj8i3qjoyc"
+    }
+  }
+}
