+++ app/build.gradle
@@ -28,16 +28,17 @@ android {
 }
 
 dependencies {
-    androidTestImplementation 'androidx.test.espresso:espresso-contrib:3.5.0-alpha05'
-    def room_version = "2.4.2"
-    implementation "androidx.room:room-runtime:$room_version"
-    annotationProcessor "androidx.room:room-compiler:$room_version"
     implementation 'androidx.appcompat:appcompat:1.4.1'
     implementation 'com.google.android.material:material:1.5.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.espresso:espresso-contrib:3.5.0-alpha05'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0-alpha05'
     androidTestImplementation 'androidx.test:rules:1.4.1-alpha05'
+
+    // AWS Amplify Dependencies
+    implementation 'com.amplifyframework:aws-api:1.33.0'
+    implementation 'com.amplifyframework:aws-datastore:1.33.0'
 }
