+++ app/src/main/java/com/kevdev/taskmaster/activities/AddTask.java
@@ -1,8 +1,5 @@
 package com.kevdev.taskmaster.activities;
 
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.room.Room;
-
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
@@ -10,21 +7,19 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
 
-import com.kevdev.taskmaster.Model.Task;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.kevdev.taskmaster.enums.State;
+import com.kevdev.taskmaster.model.Task;
 import com.kevdev.taskmaster.R;
-import com.kevdev.taskmaster.database.TaskmasterDatabase;
-import com.kevdev.taskmaster.database.TaskmasterDatabase_Impl;
 
 public class AddTask extends AppCompatActivity {
-
-    TaskmasterDatabase taskmasterDatabase;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_add_task);
 
-        taskmasterDatabase = Room.databaseBuilder(getApplicationContext(), TaskmasterDatabase.class, "task_master").allowMainThreadQueries().fallbackToDestructiveMigration().build();
         Button submitButton = (Button) findViewById(R.id.submitNewTaskButton);
         TextView confirmation = (TextView) findViewById(R.id.addTaskTextView);
 
@@ -36,8 +31,9 @@ public class AddTask extends AppCompatActivity {
                 String taskTitle = taskTitleEditText.getText().toString();
                 EditText taskBodyEditText = (EditText) findViewById(R.id.taskBody);
                 String taskBody = taskBodyEditText.getText().toString();
-                Task newTask = new Task(taskTitle, taskBody, Task.State.NEW);
-                taskmasterDatabase.taskDAO().insertATask(newTask);
+                Task newTask = new Task(taskTitle, taskBody, State.NEW);
+                //TODO convert to Dynamo query
+//                taskmasterDatabase.taskDAO().insertATask(newTask);
                 Intent goToMainActivity = new Intent(AddTask.this, MainActivity.class);
                 startActivity(goToMainActivity);
             }
