+++ build.gradle
@@ -1,4 +1,14 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
+
+buildscript {
+    dependencies {
+        classpath 'com.android.tools.build:gradle:4.1.2'
+
+        // Add this line into `dependencies` in `buildscript`
+        classpath 'com.amplifyframework:amplify-tools-gradle-plugin:1.0.2'
+    }
+}
+
 plugins {
     id 'com.android.application' version '7.1.1' apply false
     id 'com.android.library' version '7.1.1' apply false
@@ -7,3 +17,5 @@ plugins {
 task clean(type: Delete) {
     delete rootProject.buildDir
 }
+
+apply plugin: 'com.amplifyframework.amplifytools'
